Fix Upkeep thread deadlock on start-up.
When the servlet started, the Upkeep thread would get stuck at the line
mp = pool.take() until the first time a process was created. This
happens before the upkeep thread creates any new processes that are
necessary. Therefore, it was stuck until something else (a POST request)
created a process for the first tim.

The solution is to change the Upkeep thread to use the non-blocking
pool.poll().