Extend the servlet to run multiple versios of maxima.
This is a huge rewrite so that one Maxima server can run serveral
different copies of the STACK library code.

The main reason to do this is to make it easier to upgrade STACK.
Before you upgrade STACK in Moodle, you can copy the new Maxima code
onto the server, and test that it works. Then, when you upgrade Moodle,
it will automatically start using the new version. Later, you can remove
the old version of the Maxima code from the server.